add setState, processResponse to state.js, getVideoId to utils.js
